#91
Refactorización de entidades, pruebas y migraciones

Se realizaron los siguientes cambios principales:
- Se añadió un constructor en `ComprarBonoBocadilloDetalle`.
- Se eliminó `ProductoCompra` y se añadió `ApplicationUser` en el contexto.
- Se cambió el tipo de `PVP` de `float` a `double` en varias clases.
- Se reestructuraron las pruebas unitarias para simplificación y claridad.
- Se eliminaron migraciones antiguas y se añadieron nuevas con tablas actualizadas.
- Se añadieron validaciones adicionales en las pruebas de `ComprarBonoPost`.

+++ src/AppForSEII2526.API/Models/BonoBocadillo.cs
@@ -7,7 +7,7 @@ namespace AppForSEII2526.API.Models
         {
         }
 
-        public BonoBocadillo(int bonoId, int cantidadDisponible, int nBocadillos, string nombre, float pVP, TipoBocadillo tipoBocadillos, List<BonosComprados> bonosComprados)
+        public BonoBocadillo(int bonoId, int cantidadDisponible, int nBocadillos, string nombre, double pVP, TipoBocadillo tipoBocadillos, List<BonosComprados> bonosComprados)
         {
             BonoId = bonoId;
             this.cantidadDisponible = cantidadDisponible;
@@ -27,7 +27,7 @@ namespace AppForSEII2526.API.Models
         [Required]
         public string nombre { get; set;}
         [Required]
-        public float PVP { get; set; } 
+        public double PVP { get; set; } 
 
         public TipoBocadillo tipoBocadillos { get; set; }
         public List<BonosComprados> bonosComprados { get; set; } = new List<BonosComprados>();
